chore: rename model to models
